% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations_info.R
\name{citations_info}
\alias{citations_info}
\title{Get citations info}
\usage{
citations_info(id)
}
\arguments{
\item{id}{A character vector with the Google Scholar ID}
}
\value{
A data frame with the ID, name, total citations and h-index
}
\description{
Retrieve the total number of citations and h-index for a given Google Scholar ID.
}
\examples{
ids <- c("GNJ8XGQAAAAJ","Wy6KhVIAAAAJ")

citations_info(ids)

}
