% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations_year.R
\name{citations_year}
\alias{citations_year}
\title{citations_year}
\usage{
citations_year(id)
}
\arguments{
\item{id}{character string of the researcher's Google Scholar ID}
}
\value{
a data frame with the citation history
}
\description{
Get the citation history of a researcher from Google Scholar
}
