% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentage_scopus.R
\name{get_percentage_scopus}
\alias{get_percentage_scopus}
\title{Get the Scopus percentage}
\usage{
get_percentage_scopus(
  journal_id = NULL,
  journal_name = NULL,
  year = 2021,
  best_rank = TRUE
)
}
\arguments{
\item{journal_id}{The ID of the journal.}

\item{journal_name}{The name of the journal.}

\item{year}{The year to get the percentage from.}
}
\value{
A tibble with the percentage of the journal in Scopus.
}
\description{
Get the Scopus percentage of a given journal.
}
\examples{
get_percentage_scopus(journal_id = "5200153106", year = 2021)

}
