% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_gscholar_nupelia.R
\docType{data}
\name{ids_gscholar_nupelia}
\alias{ids_gscholar_nupelia}
\title{IDs for Google Scholar of nupelia workers}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\usage{
ids_gscholar_nupelia
}
\value{
A data frame with the ID and name for nupelia workers
}
\description{
IDs for Google Scholar of nupelia workers
}
\keyword{datasets}
