% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peld.R
\name{peld_abrevia_especie}
\alias{peld_abrevia_especie}
\title{PELD: Abrevia espécie}
\usage{
peld_abrevia_especie(dados)
}
\arguments{
\item{dados}{Tabela de dados do PELD}
}
\description{
PELD: Abrevia espécie
}
