% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peld.R
\name{peld_abund_rel_subsistema}
\alias{peld_abund_rel_subsistema}
\title{PELD: Abundância relativa por subsistema}
\usage{
peld_abund_rel_subsistema(
  dados,
  subsistema = "geral",
  porcentagem_corte = 2,
  especie_var = "especie",
  ivinhema_var = "ivi",
  baia_var = "bai",
  parana_var = "par"
)
}
\arguments{
\item{dados}{Tabela de dados do PELD}

\item{subsistema}{Subsistema indicado. Pode ser "ivinhema", "baia" ou "parana". "geral" por padrao.}

\item{porcentagem_corte}{A porcentagem da abundância relativa utilizada como corte para espécies. Espécies com porcentagem menor serão agrupadas em "Outros". 2 por padrão.}

\item{especie_var}{Nome da variável referente a "Espécie". "especie" por padrao.}

\item{ivinhema_var}{Nome da variável referente a "ivinhema". "ivinhema" por padrao.}

\item{baia_var}{Nome da variável referente a "baia". "baia" por padrao.}

\item{parana_var}{Nome da variável referente a "parana". "parana" por padrao.}
}
\description{
PELD: Abundância relativa por subsistema
}
