% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peld.R
\name{peld_equitabilidade}
\alias{peld_equitabilidade}
\title{PELD: Equitabilidade de Pielou (J)}
\usage{
peld_equitabilidade(
  dados,
  subsistema = "geral",
  especie_var = "especie",
  ivinhema_var = "ivi",
  baia_var = "bai",
  parana_var = "par"
)

peld_equitabilidade(
  dados,
  subsistema = "geral",
  especie_var = "especie",
  ivinhema_var = "ivi",
  baia_var = "bai",
  parana_var = "par"
)
}
\arguments{
\item{dados}{Tabela de dados do PELD}

\item{subsistema}{Subsistema indicado. Pode ser "ivinhema", "baia" ou "parana". "geral" por padrao.}

\item{especie_var}{Nome da variável referente a "Espécie". "especie" por padrao.}

\item{ivinhema_var}{Nome da variável referente a "ivinhema". "ivinhema" por padrao.}

\item{baia_var}{Nome da variável referente a "baia". "baia" por padrao.}

\item{parana_var}{Nome da variável referente a "parana". "parana" por padrao.}

\item{ambiente}{Ambiente indicado. Pode ser "geral", "lab", "lfe" ou "rio". "geral" por padrao.}

\item{lab_var}{Nome da variável referente a "Lagoas abertas". "lab" por padrao.}

\item{lfe_var}{Nome da variável referente a "Lagoas fechadas". "lfe" por padrao.}

\item{rio_var}{Nome da variável referente a "Rios". "rio" por padrao.}
}
\description{
PELD: Equitabilidade de Pielou (J)

PELD: Equitabilidade de Pielou (J) por subsistema
}
