% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peld.R
\name{peld_riqueza_subsistema}
\alias{peld_riqueza_subsistema}
\title{PELD: Riqueza por subsistema}
\usage{
peld_riqueza_subsistema(
  dados,
  subsistema = "geral",
  especie_var = "especie",
  ivinhema_var = "ivi",
  baia_var = "bai",
  parana_var = "par"
)
}
\arguments{
\item{dados}{Tabela de dados do PELD}

\item{subsistema}{Subsistema indicado. Pode ser "ivinhema", "baia" ou "parana". "geral" por padrao.}

\item{especie_var}{Nome da variável referente a "Espécie". "especie" por padrao.}

\item{ivinhema_var}{Nome da variável referente a "ivinhema". "ivinhema" por padrao.}

\item{baia_var}{Nome da variável referente a "baia". "baia" por padrao.}

\item{parana_var}{Nome da variável referente a "parana". "parana" por padrao.}
}
\description{
PELD: Riqueza por subsistema
}
