% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_scopus.R
\docType{data}
\name{tabela_scopus}
\alias{tabela_scopus}
\title{Tabela de revistas da Scopus}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43400 rows and 2 columns.
}
\usage{
tabela_scopus
}
\description{
Tabela de revistas da Scopus
}
\keyword{datasets}
