% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peld.R
\name{peld_abund_rel}
\alias{peld_abund_rel}
\title{PELD: Abundância relativa}
\usage{
peld_abund_rel(
  dados,
  subsistema = "geral",
  especie_var = especie,
  lab_var = lab,
  lfe_var = lfe,
  rio_var = rio
)
}
\arguments{
\item{dados}{Tabela de dados do PELD}

\item{subsistema}{Subsistema indicado. Pode ser "geral", "lab", "lfe" ou "rio". "geral" por padrao.}

\item{especie_var}{Nome da variável referente a "Espécie". "especie" por padrao.}

\item{lab_var}{Nome da variável referente a "Lagoas abertas". "lab" por padrao.}

\item{lfe_var}{Nome da variável referente a "Lagoas fechadas". "lfe" por padrao.}

\item{rio_var}{Nome da variável referente a "Rios". "rio" por padrao.}
}
\value{

}
\description{
PELD: Abundância relativa
}
