% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peld.R
\name{peld_equitabilidade}
\alias{peld_equitabilidade}
\title{PELD: Equitabilidade de Pielou (J)}
\usage{
peld_equitabilidade(
  dados,
  ambiente = "geral",
  especie_var = especie,
  lab_var = lab,
  lfe_var = lfe,
  rio_var = rio
)
}
\arguments{
\item{dados}{Tabela de dados do PELD}

\item{ambiente}{Ambiente indicado. Pode ser "geral", "lab", "lfe" ou "rio". "geral" por padrao.}

\item{especie_var}{Nome da variável referente a "Espécie". "especie" por padrao.}

\item{lab_var}{Nome da variável referente a "Lagoas abertas". "lab" por padrao.}

\item{lfe_var}{Nome da variável referente a "Lagoas fechadas". "lfe" por padrao.}

\item{rio_var}{Nome da variável referente a "Rios". "rio" por padrao.}
}
\description{
PELD: Equitabilidade de Pielou (J)
}
